+++ packages/docs/src/lib/liveCode/dynamicImports.ts
@@ -2,10 +2,10 @@
 /* auto-generated file */
 import { ImportMapping } from "@/lib/liveCode/types";
 import { lazy } from "react";
-import { NavigationItem as I33 } from "@mittwald/flow-react-components/Navigation";
-import { ListItemView as I44 } from "@mittwald/flow-react-components/List";
-import { ListStaticData as I45 } from "@mittwald/flow-react-components/List";
-import { users as I46 } from "@/content/02-components/structure/list/examples/userApi";
+import { NavigationItem as I30 } from "@mittwald/flow-react-components/Navigation";
+import { ListItemView as I43 } from "@mittwald/flow-react-components/List";
+import { ListStaticData as I44 } from "@mittwald/flow-react-components/List";
+import { users as I45 } from "@/content/02-components/structure/list/examples/userApi";
 
 export const liveCodeEditorGlobalImports: ImportMapping = {
   "ExampleSvg:@/content/01-getting-started/stylesheet/examples/components/ExampleSvg": lazy(() => import("@/content/01-getting-started/stylesheet/examples/components/ExampleSvg").then(module => ({ default: module.ExampleSvg } ))),
@@ -14,6 +14,7 @@ export const liveCodeEditorGlobalImports: ImportMapping = {
 "Text:@mittwald/flow-react-components/Text": lazy(() => import("@mittwald/flow-react-components/Text").then(module => ({ default: module.Text } ))),
 "IconChevronDown:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconChevronDown } ))),
 "CopyButton:@mittwald/flow-react-components/CopyButton": lazy(() => import("@mittwald/flow-react-components/CopyButton").then(module => ({ default: module.CopyButton } ))),
+"ButtonGroup:@mittwald/flow-react-components/ButtonGroup": lazy(() => import("@mittwald/flow-react-components/ButtonGroup").then(module => ({ default: module.ButtonGroup } ))),
 "ColumnLayout:@mittwald/flow-react-components/ColumnLayout": lazy(() => import("@mittwald/flow-react-components/ColumnLayout").then(module => ({ default: module.ColumnLayout } ))),
 "TextField:@mittwald/flow-react-components/TextField": lazy(() => import("@mittwald/flow-react-components/TextField").then(module => ({ default: module.TextField } ))),
 "Label:@mittwald/flow-react-components/Label": lazy(() => import("@mittwald/flow-react-components/Label").then(module => ({ default: module.Label } ))),
@@ -33,17 +34,15 @@ export const liveCodeEditorGlobalImports: ImportMapping = {
 "NumberField:@mittwald/flow-react-components/NumberField": lazy(() => import("@mittwald/flow-react-components/NumberField").then(module => ({ default: module.NumberField } ))),
 "RadioGroup:@mittwald/flow-react-components/RadioGroup": lazy(() => import("@mittwald/flow-react-components/RadioGroup").then(module => ({ default: module.RadioGroup } ))),
 "Radio:@mittwald/flow-react-components/RadioGroup": lazy(() => import("@mittwald/flow-react-components/RadioGroup").then(module => ({ default: module.Radio } ))),
-"Select:@mittwald/flow-react-components/Select": lazy(() => import("@mittwald/flow-react-components/Select").then(module => ({ default: module.Select } ))),
-"Option:@mittwald/flow-react-components/Select": lazy(() => import("@mittwald/flow-react-components/Select").then(module => ({ default: module.Option } ))),
-"Options:@mittwald/flow-react-components/Select": lazy(() => import("@mittwald/flow-react-components/Select").then(module => ({ default: module.Options } ))),
-"React:react": lazy(() => import("react").then(module => ({ default: module.React } ))),
 "Switch:@mittwald/flow-react-components/Switch": lazy(() => import("@mittwald/flow-react-components/Switch").then(module => ({ default: module.Switch } ))),
 "FieldDescription:@mittwald/flow-react-components/FieldDescription": lazy(() => import("@mittwald/flow-react-components/FieldDescription").then(module => ({ default: module.FieldDescription } ))),
 "TextArea:@mittwald/flow-react-components/TextArea": lazy(() => import("@mittwald/flow-react-components/TextArea").then(module => ({ default: module.TextArea } ))),
 "Navigation:@mittwald/flow-react-components/Navigation": lazy(() => import("@mittwald/flow-react-components/Navigation").then(module => ({ default: module.Navigation } ))),
-"NavigationItem:@mittwald/flow-react-components/Navigation": I33,
+"NavigationItem:@mittwald/flow-react-components/Navigation": I30,
 "IconProject:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconProject } ))),
 "IconServer:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconServer } ))),
+"Modal:@mittwald/flow-react-components/Modal": lazy(() => import("@mittwald/flow-react-components/Modal").then(module => ({ default: module.Modal } ))),
+"ModalTrigger:@mittwald/flow-react-components/Modal": lazy(() => import("@mittwald/flow-react-components/Modal").then(module => ({ default: module.ModalTrigger } ))),
 "Tooltip:@mittwald/flow-react-components/Tooltip": lazy(() => import("@mittwald/flow-react-components/Tooltip").then(module => ({ default: module.Tooltip } ))),
 "TooltipTrigger:@mittwald/flow-react-components/Tooltip": lazy(() => import("@mittwald/flow-react-components/Tooltip").then(module => ({ default: module.TooltipTrigger } ))),
 "IconCopy:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconCopy } ))),
@@ -52,9 +51,9 @@ export const liveCodeEditorGlobalImports: ImportMapping = {
 "StatusIcon:@mittwald/flow-react-components/StatusIcon": lazy(() => import("@mittwald/flow-react-components/StatusIcon").then(module => ({ default: module.StatusIcon } ))),
 "LayoutCard:@mittwald/flow-react-components/LayoutCard": lazy(() => import("@mittwald/flow-react-components/LayoutCard").then(module => ({ default: module.LayoutCard } ))),
 "List:@mittwald/flow-react-components/List": lazy(() => import("@mittwald/flow-react-components/List").then(module => ({ default: module.List } ))),
-"ListItemView:@mittwald/flow-react-components/List": I44,
-"ListStaticData:@mittwald/flow-react-components/List": I45,
-"users:@/content/02-components/structure/list/examples/userApi": I46,
+"ListItemView:@mittwald/flow-react-components/List": I43,
+"ListStaticData:@mittwald/flow-react-components/List": I44,
+"users:@/content/02-components/structure/list/examples/userApi": I45,
 "Section:@mittwald/flow-react-components/Section": lazy(() => import("@mittwald/flow-react-components/Section").then(module => ({ default: module.Section } ))),
 "Header:@mittwald/flow-react-components/Header": lazy(() => import("@mittwald/flow-react-components/Header").then(module => ({ default: module.Header } ))),
 };