+++ packages/components/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.51](https://github.com/mittwald/flow/compare/0.1.0-alpha.50...0.1.0-alpha.51) (2024-03-22)
+
+### Features
+
+- move Tunnel into own package
+  ([5a2ba52](https://github.com/mittwald/flow/commit/5a2ba5206a229df93f9c59493fafe2b797798e1d))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.50](https://github.com/mittwald/flow/compare/0.1.0-alpha.49...0.1.0-alpha.50) (2024-03-22)
+
+### Features
+
+- add Modal component
+  ([d9a47ea](https://github.com/mittwald/flow/commit/d9a47eae58db9c7a8b1b6868da7246e1eded2a16))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.49](https://github.com/mittwald/flow/compare/0.1.0-alpha.48...0.1.0-alpha.49) (2024-03-22)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.48](https://github.com/mittwald/flow/compare/0.1.0-alpha.47...0.1.0-alpha.48) (2024-03-21)
+
+### Features
+
+- add Tunnel component
+  ([ffca8e1](https://github.com/mittwald/flow/commit/ffca8e17d32ef8fd452200ece54a59b8e5409861))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.47](https://github.com/mittwald/flow/compare/0.1.0-alpha.46...0.1.0-alpha.47) (2024-03-21)
 
 ### Bug Fixes